requirements.txt.txt
openai==1.30.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4

